reslove #2 refactor:domain -> entity 패키지명 변경
