+++ src/main/java/iceapple/placeservice/repository/RoomRepository.java
@@ -1,11 +1,9 @@
 package iceapple.placeservice.repository;
 
-import iceapple.placeservice.domain.Room;
-import iceapple.placeservice.dto.response.RoomTimeCountResponse;
+import iceapple.placeservice.entity.Room;
 import iceapple.placeservice.util.TimeCount;
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Timer;
 
 public interface RoomRepository {
     List<Room> findAll();
