+++ src/main/java/iceapple/placeservice/service/ReservationService.java
@@ -1,6 +1,6 @@
 package iceapple.placeservice.service;
 
-import iceapple.placeservice.domain.Reservation;
+import iceapple.placeservice.entity.Reservation;
 import iceapple.placeservice.dto.request.ReservationRequest;
 import iceapple.placeservice.repository.ReservationRepository;
 import java.util.List;
