+++ src/main/java/iceapple/placeservice/service/RoomService.java
@@ -1,12 +1,11 @@
 package iceapple.placeservice.service;
 
-import iceapple.placeservice.domain.Room;
+import iceapple.placeservice.entity.Room;
 import iceapple.placeservice.dto.response.RoomTimeCountResponse;
 import iceapple.placeservice.repository.RoomRepository;
 import iceapple.placeservice.util.TimeCount;
 import java.time.LocalDate;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
 
 public class RoomService {
